pyproject.toml
[project]
name = "blood-vessel-segmentation-unet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "gdown>=5.2.0",
    "glob2>=0.7",
    "imageio>=2.37.0",
    "numpy>=2.3.1",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "streamlit>=1.47.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]
